refactor: Change deposit points and weight data type to float
Updates the data type for deposit points and weight from integer to float across the navigation graph and related screens.

- In `AuthenticatedNavHost`, the `navArgument` types for the `SuccessDeposit` route are changed from `NavType.IntType` to `NavType.FloatType`.
- In `SuccessDepositScreen`, the `points` and `weight` parameters are updated from `Int` to `Float` to correctly handle the floating-point values passed through navigation.

+++ app/src/main/java/com/ahargunyllib/growth/presentation/ui/navigation/nav_host/AuthenticatedNavHost.kt
@@ -93,8 +93,8 @@ fun AuthenticatedNavHost(rootNavController: NavController) {
                         navArgument("weight") { type = NavType.FloatType }
                     ),
                     content = { backStackEntry ->
-                        val points = backStackEntry.arguments?.getInt("points") ?: 0F
-                        val weight = backStackEntry.arguments?.getInt("weight") ?: 0F
+                        val points = backStackEntry.arguments?.getFloat("points") ?: 0F
+                        val weight = backStackEntry.arguments?.getFloat("weight") ?: 0F
                         SuccessDepositScreen(
                             authenticatedNavController = authenticatedNavController,
                             points = points,
